==> R_Scripts/99_Source_File.R <==
rm(list=ls())
setwd("/Users/howquez/Documents/002_UNI/UCPH/016_Master Thesis/05_Data")

source("03_RCodes/01_Read_XML_Data.R")
source("03_RCodes/02_Data_Manipulation.R")
source("03_RCodes/03_Data_Analysis.R")
source("03_RCodes/04_Graphs&Tables.R")
source("03_RCodes/06_SubSample.R")
source("03_RCodes/07_Shiny.R")
